backend/requirements.txt
fastapi
uvicorn
python-multipart
openai-whisper
torch
torchaudio
transformers
f5-tts
ffmpeg-python
numpy
loguru
